+++ internals/native/multiply_parallel_worker_pool.go
@@ -20,15 +20,13 @@ func (m *Matrix) MultiplyParallelWorkerPool(b *Matrix) (*Matrix, error) {
 		return nil, fmt.Errorf("incompattible matrix dimensions for multiplication")
 	}
 
-	result := make([][]float32, len(*m))
+	result := make([][]float64, len(*m))
 	for i := range result {
-		result[i] = make([]float32, len((*b)[0]))
+		result[i] = make([]float64, len((*b)[0]))
 	}
 
 	numWorkers := runtime.NumCPU()
-
 	log.Printf("Number of CPU's available: %d\n", numWorkers)
-
 	jobs := make(chan int, len(*m))
 
 	var wg sync.WaitGroup
@@ -39,7 +37,7 @@ func (m *Matrix) MultiplyParallelWorkerPool(b *Matrix) (*Matrix, error) {
 			defer wg.Done()
 			for rowIdx := range jobs {
 				for j := range (*b)[0] {
-					var sum float32 = 0.0
+					var sum float64 = 0.0
 					for k := range *b {
 						sum += (*m)[rowIdx][k] * (*b)[k][j]
 					}
